moved and fixed some stuff
